// src/app/grass/Models/AllotmentMaster.ts
export class AllotmentMaster {
  ID: any = 0;
  // QUARTER_ID : string = '';
  SUPERANNUATION_DATE: any;
  PRESENT_RESENDENTIAL_ADDRESS: string = '';
  PRESENT_OFFICE_ADDRESS: string = '';
  // DESIGNATION: any;
  DESIGNATION_ID: any;
  OFFICER_CAST: string = '';
  DATE_JOINING: any;
  CAST: any;
  JOINING_LETTER: any;
  JOINING_ALLOTMENT_YEAR: boolean = false;
  JOINING_ALLOTEMENT_DETAILS: string = '';
  JOINING_ALLOTMENT_LETTER: any;
  BASIC_PAY: string = '';
  GRADE_PAY: string = '';
  PAY_LEVEL: string = '';
  GRADE_PAY_DRAWN_DATE: any;
  GRADE_PAY_DRAWN_LETTER: any;
  APPLICATION_FORM: any;
  IS_EMOLUMENTS_DRAWN: boolean = false;
  EMOLUMENTS_DRAWN: string = '';
  ACCOMMODATION_TYPE: string = '';
  ACCOMMODATION_SENIORITY_DATE: any;
  DATE_MIN_PAY_GOV: string = '';
  EMPLOYEE_ID: any;
  STATUS: any;
  GRASS_INSPECTOR_ID: string = '';
  FLAT_PREFERENCE: any = [];
  PAY_SLIP: any;
  I_CARD: any;
  REMARK: string = '';
  STEP_NO: number = 0;
  APP_APPROVED_DATETIME: any;
  APP_SUBMIT_DATETIME: any;
  APP_INFO_DATETIME: any;
  APP_PRINT_DATETIME: any;
  APPLICATION_DATETIME: any;
  DATE_CONDITION: any;
  CAST_CERTIFICATE: any = null;
  DEMO_GRADE_PAY_ID: any;
  IS_SPOUSE_WORKING_IN_GOV: any;
  SPOUSE_EMP_CODE: any;
  SPOUSE_NAME: any;
  MARRIAGE_CERTIFICATE: any;
  PRESENT_PAY_LEVEL_DATE: any;
  IS_PRESENT_PAY_LEVEL_CHANGED: any = 0;
}