// src/app/grass/Models/ObjectionMaster.ts
export class ObjectionMaster{
    EMPLOYEE_ID:any
    SENIORITY_LIST_ID:any
    REASON: string = '';
    RESOLVED_BY_ID : any;
    ATTACHMENT1: any;
    REMARK: string = '';
    FLAT_REQUEST_ID:any;
    ATTACHMENT2: any ;
    REQUEST_DATETIME: any;
    RESOLVED_DATETIME: any;
    STATUS : string = '';
}